fix trivial lgtm errors

+++ src/autopas/containers/verletClusterLists/VerletClusterLists.h
@@ -302,7 +302,7 @@ class VerletClusterLists : public ParticleContainer<Particle, FullParticleCell<P
                   float jBBoxBot = jClusterStart->getR()[2];
                   float jBBoxTop = (jClusterStart + (_clusterSize - 1))->getR()[2];
 
-                  float distZ = bboxDistance(iBBoxBot, iBBoxTop, jBBoxBot, jBBoxTop);
+                  double distZ = bboxDistance(iBBoxBot, iBBoxTop, jBBoxBot, jBBoxTop);
                   if (distXYsqr + distZ * distZ <= _cutoffSqr) {
                     iClusterVerlet.push_back(jClusterStart);
                   }
@@ -314,7 +314,7 @@ class VerletClusterLists : public ParticleContainer<Particle, FullParticleCell<P
                   float jBBoxBot = jClusterStart->getR()[2];
                   float jBBoxTop = (jClusterStart + (jRest - 1))->getR()[2];
 
-                  float distZ = bboxDistance(iBBoxBot, iBBoxTop, jBBoxBot, jBBoxTop);
+                  double distZ = bboxDistance(iBBoxBot, iBBoxTop, jBBoxBot, jBBoxTop);
                   if (distXYsqr + distZ * distZ <= _cutoffSqr) {
                     iClusterVerlet.push_back(jClusterStart);
                   }
@@ -332,7 +332,7 @@ class VerletClusterLists : public ParticleContainer<Particle, FullParticleCell<P
                   float jBBoxBot = jClusterStart->getR()[2];
                   float jBBoxTop = (jClusterStart + (_clusterSize - 1))->getR()[2];
 
-                  float distZ = bboxDistance(iBBoxBot, iBBoxTop, jBBoxBot, jBBoxTop);
+                  double distZ = bboxDistance(iBBoxBot, iBBoxTop, jBBoxBot, jBBoxTop);
                   if (distXYsqr + distZ * distZ <= _cutoffSqr) {
                     iClusterVerlet.push_back(jClusterStart);
                   }
@@ -344,7 +344,7 @@ class VerletClusterLists : public ParticleContainer<Particle, FullParticleCell<P
                   float jBBoxBot = jClusterStart->getR()[2];
                   float jBBoxTop = (jClusterStart + (jRest - 1))->getR()[2];
 
-                  float distZ = bboxDistance(iBBoxBot, iBBoxTop, jBBoxBot, jBBoxTop);
+                  double distZ = bboxDistance(iBBoxBot, iBBoxTop, jBBoxBot, jBBoxTop);
                   if (distXYsqr + distZ * distZ <= _cutoffSqr) {
                     iClusterVerlet.push_back(jClusterStart);
                   }
